src/main/java/com/zuhlke/paint/ascii/figure/Point.java
package com.zuhlke.paint.ascii.figure;

import java.util.Objects;

/**
 * The {@code Point} class represents a two dimensional coordinate (x,y).
 * It is modeled as a value object.
 **
 */
public class Point {

    private int x;
    private int y;

    private Point(int x, int y) {
        this.x = x;
        this.y = y;
    }

    public static Point create(int x, int y) {
        return new Point(x, y);
    }

    public int getX() {
        return x;
    }

    public int getY() {
        return y;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Point point = (Point) o;
        return x == point.x &&
                y == point.y;
    }

    @Override
    public int hashCode() {
        return Objects.hash(x, y);
    }


}
